chore(website): upgrade docusaurus to v3
Closes #62
