+++ packages/website/docs/components/Balloon.mdx
@@ -37,7 +37,7 @@ A tiny wrapper around [Link](./Link) component, to be used for actions in Balloo
 
 :::info Note
 How balloon notifications are stacked and positioned in bottom right, and an imperative API for triggering notifications
-is outside the scope of what this component is about. Use [BalloonManager](#baloonmanager) and
+is outside the scope of what this component is about. Use [BalloonManager](#balloonmanager) and
 [useBalloonManager](#useballoonmanager) for that.
 :::
