+++ packages/website/docs/components/SpeedSearch.mdx
@@ -18,14 +18,14 @@ original implementation.
 - Highlighting matches in items
 - Showing search query as an overlay at the top left, with visual feedback about search result.
 - [Low level](#low-level-speed-search-api) API for implementing speed search for any UI component.
-- [Higher level API](#collcetion-speed-search-API) for integrating speed search with Collection components, such as `List`, `Tree`, etc.
+- [Higher level API](#collection-speed-search-api) for integrating speed search with Collection components, such as `List`, `Tree`, etc.
 
 ## Integration with other components
 
 Normally, you don't need to use speed search hooks and components directly. Collection component like `List` or `Tree`
 come with a version with built-in support for speed search.
 
-See [SpeedSearchList](./list#speedsearchlist), [SpeedSearchTree](./Tree#speedsearchtree).
+See [SpeedSearchList](./List#speedsearchlist), [SpeedSearchTree](./Tree#speedsearchtree).
 
 ## Advanced use
