+++ packages/website/docs/components/ToolWindows.mdx
@@ -139,7 +139,7 @@ render(<App />);
 ### MultiViewToolWindow
 
 Built on top of [DefaultToolWindow](#defaulttoolwindow), extends it to support multiple views within the same tool
-window, switched by either [Tabs](./tabs#tool-window-theme) or a dropdown.
+window, switched by either [Tabs](./Tabs#tool-window-theme) or a dropdown.
 
 :::warning tab grouping
 Grouping tabs in a dropdown is not implemented at the moment, and passing `true` to `groupTabs` will throw a warning.
