+++ packages/website/docs/guides/ActionSystem.mdx
@@ -105,7 +105,7 @@ An ActionGroup is a special type of action with a list of children actions. An a
 
 ### Default action group
 
-A default implementation of action group is available, which when performed, opens a [Popup](../components/popup),
+A default implementation of action group is available, which when performed, opens a [Popup](../components/Popup),
 presenting the children actions in a [SpeedSearchMenu](../components/Menu#speedsearchmenu). The default action
 group requires [PopupManager](../components/Popup#popupmanager), for opening the popup.
