+++ packages/website/package.json
@@ -18,28 +18,31 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.1",
-    "@docusaurus/module-type-aliases": "^2.4.1",
-    "@docusaurus/preset-classic": "^2.4.1",
-    "@docusaurus/theme-live-codeblock": "^2.4.1",
+    "@docusaurus/core": "^3.7.0",
+    "@docusaurus/faster": "^3.7.0",
+    "@docusaurus/preset-classic": "^3.7.0",
+    "@docusaurus/theme-live-codeblock": "^3.7.0",
     "@intellij-platform/core": "workspace:^",
-    "@mdx-js/react": "^1.6.21",
+    "@mdx-js/react": "^3.1.0",
     "@svgr/webpack": "^5.5.0",
     "clsx": "^1.1.1",
     "file-loader": "^6.2.0",
     "markdown-to-jsx": "^7.1.7",
-    "prism-react-renderer": "^1.2.1",
+    "prism-react-renderer": "^2.4.1",
     "react": "*",
     "react-dom": "*",
     "styled-components": "^5",
     "url-loader": "^4.1.1"
   },
   "devDependencies": {
+    "@docusaurus/module-type-aliases": "^3.7.0",
+    "@docusaurus/tsconfig": "^3.7.0",
+    "@docusaurus/types": "^3.7.0",
     "@microsoft/api-extractor-model": "^7.18.0",
     "@react-aria/ssr": "^3.9.5",
-    "@tsconfig/docusaurus": "^1.0.6",
     "babel-plugin-react-docgen": "^4.2.1",
     "jsdoc-parse-plus": "^1.3.0",
+    "typescript": "*",
     "webpack": "^5.97.1"
   },
   "browserslist": {
